pyproject.toml
[project]
name = "conversational-search"
version = "0.1.0"
description = "Conversational search engine with citations and comprehensive evals"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "braintrust[cli]",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
    "autoevals",
    "ipython>=9.7.0",
    "rich>=13.0.0",
    "langchain-openai>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
